refactor: UserGroup Domain Root로 변경
