+++ src/main/java/com/gloddy/server/article/domain/service/ArticleDeletePolicy.java
@@ -2,7 +2,7 @@ package com.gloddy.server.article.domain.service;
 
 import com.gloddy.server.article.domain.Article;
 import com.gloddy.server.article.exception.NoAuthorizedArticleDeleteException;
-import com.gloddy.server.group.domain.UserGroup;
+import com.gloddy.server.user_group.domain.UserGroup;
 import org.springframework.stereotype.Component;
 
 @Component
