+++ src/main/java/com/gloddy/server/article/domain/service/NoticeArticleCreatePolicy.java
@@ -1,7 +1,7 @@
 package com.gloddy.server.article.domain.service;
 
 import com.gloddy.server.article.exception.NoAuthorizedNoticeArticleException;
-import com.gloddy.server.group.domain.UserGroup;
+import com.gloddy.server.user_group.domain.UserGroup;
 import org.springframework.stereotype.Component;
 
 @Component
