+++ src/main/java/com/gloddy/server/estimate/domain/vo/strategy/KindPraiseStrategy.java
@@ -1,6 +1,6 @@
 package com.gloddy.server.estimate.domain.vo.strategy;
 
-import com.gloddy.server.group.domain.UserGroup;
+import com.gloddy.server.user_group.domain.UserGroup;
 
 public class KindPraiseStrategy implements PraiseStrategy {
     @Override
