+++ src/main/java/com/gloddy/server/group/application/UserGroupFindService.java
@@ -1,7 +1,7 @@
 package com.gloddy.server.group.application;
 
-import com.gloddy.server.group.domain.UserGroup;
-import com.gloddy.server.group.infra.repository.UserGroupJpaRepository;
+import com.gloddy.server.user_group.domain.UserGroup;
+import com.gloddy.server.user_group.infra.repository.UserGroupJpaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
