+++ src/main/java/com/gloddy/server/user_group/domain/handler/UserGroupQueryHandler.java
@@ -1,6 +1,6 @@
-package com.gloddy.server.group.domain.handler;
+package com.gloddy.server.user_group.domain.handler;
 
-import com.gloddy.server.group.domain.UserGroup;
+import com.gloddy.server.user_group.domain.UserGroup;
 
 public interface UserGroupQueryHandler {
     UserGroup findByUserIdAndGroupId(Long userId, Long groupId);
