+++ src/main/java/com/gloddy/server/user_group/infra/repository/custom/UserGroupJpaRepositoryCustom.java
@@ -1,8 +1,8 @@
-package com.gloddy.server.group.infra.repository.custom;
+package com.gloddy.server.user_group.infra.repository.custom;
 
 import com.gloddy.server.auth.domain.User;
 import com.gloddy.server.group.domain.Group;
-import com.gloddy.server.group.domain.UserGroup;
+import com.gloddy.server.user_group.domain.UserGroup;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
