+++ src/test/java/com/gloddy/server/acceptance/reliability/UpdateReliabilityByEstimateTest.java
@@ -8,7 +8,7 @@ import com.gloddy.server.estimate.domain.vo.PraiseValue;
 import com.gloddy.server.estimate.application.mate.MateSaveService;
 import com.gloddy.server.estimate.application.praise.PraiseService;
 import com.gloddy.server.group.domain.Group;
-import com.gloddy.server.group.domain.UserGroup;
+import com.gloddy.server.user_group.domain.UserGroup;
 import com.gloddy.server.group.application.UserGroupUpdateService;
 import com.gloddy.server.reliability.domain.Reliability;
 import com.gloddy.server.reliability.domain.vo.ReliabilityLevel;
