+++ src/test/java/com/gloddy/server/common/apply/ApplyApiTest.java
@@ -6,7 +6,7 @@ import com.gloddy.server.auth.domain.User;
 import com.gloddy.server.common.BaseApiTest;
 import com.gloddy.server.group.domain.Group;
 import com.gloddy.server.group.infra.repository.GroupJpaRepository;
-import com.gloddy.server.group.infra.repository.UserGroupJpaRepository;
+import com.gloddy.server.user_group.infra.repository.UserGroupJpaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 
 public abstract class ApplyApiTest extends BaseApiTest {
