+++ src/test/java/com/gloddy/server/common/group/GroupApiTest.java
@@ -4,9 +4,9 @@ import com.gloddy.server.auth.domain.User;
 import com.gloddy.server.common.BaseApiTest;
 import com.gloddy.server.group.domain.dto.GroupRequest;
 import com.gloddy.server.group.domain.Group;
-import com.gloddy.server.group.domain.UserGroup;
+import com.gloddy.server.user_group.domain.UserGroup;
 import com.gloddy.server.group.infra.repository.GroupJpaRepository;
-import com.gloddy.server.group.infra.repository.UserGroupJpaRepository;
+import com.gloddy.server.user_group.infra.repository.UserGroupJpaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import java.time.LocalDate;
