+++ src/test/java/com/gloddy/server/common/reliability/ReliabilityApiTest.java
@@ -9,9 +9,9 @@ import com.gloddy.server.estimate.domain.vo.PraiseValue;
 import com.gloddy.server.estimate.infra.repository.MateJpaRepository;
 import com.gloddy.server.group.domain.dto.GroupRequest;
 import com.gloddy.server.group.domain.Group;
-import com.gloddy.server.group.domain.UserGroup;
+import com.gloddy.server.user_group.domain.UserGroup;
 import com.gloddy.server.group.infra.repository.GroupJpaRepository;
-import com.gloddy.server.group.infra.repository.UserGroupJpaRepository;
+import com.gloddy.server.user_group.infra.repository.UserGroupJpaRepository;
 import com.gloddy.server.reliability.domain.Reliability;
 import com.gloddy.server.reliability.domain.handler.ReliabilityQueryHandler;
 import com.gloddy.server.reliability.infra.repository.ReliabilityRepository;
