+++ src/test/java/com/gloddy/server/query/QueryTest.java
@@ -4,7 +4,7 @@ import com.gloddy.server.article.infra.repository.ArticleJpaRepository;
 import com.gloddy.server.comment.infra.repository.CommentJpaRepository;
 import com.gloddy.server.estimate.infra.repository.PraiseJpaRepository;
 import com.gloddy.server.group.infra.repository.GroupJpaRepository;
-import com.gloddy.server.group.infra.repository.UserGroupJpaRepository;
+import com.gloddy.server.user_group.infra.repository.UserGroupJpaRepository;
 import com.gloddy.server.reliability.infra.repository.ReliabilityRepository;
 import com.gloddy.server.user.infra.repository.UserJpaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
