+++ src/test/java/com/gloddy/server/query/group/FindUserGroupsToPraiseByUserIdInAndGroupIdTest.java
@@ -4,7 +4,7 @@ import com.gloddy.server.auth.domain.User;
 import com.gloddy.server.auth.domain.vo.kind.Personality;
 import com.gloddy.server.estimate.domain.Praise;
 import com.gloddy.server.group.domain.Group;
-import com.gloddy.server.group.domain.UserGroup;
+import com.gloddy.server.user_group.domain.UserGroup;
 import com.gloddy.server.query.QueryTest;
 import com.gloddy.server.reliability.domain.Reliability;
 import org.junit.jupiter.api.Test;
